+++ dart/test/contexts_test.dart
@@ -6,21 +6,21 @@ import 'dart:convert';
 
 import 'package:collection/collection.dart';
 import 'package:sentry/sentry.dart';
-import 'package:sentry/src/protocol/gpu.dart';
+import 'package:sentry/src/protocol/sentry_gpu.dart';
 import 'package:test/test.dart';
 
 void main() {
   group(Contexts, () {
     final testBootTime = DateTime.fromMicrosecondsSinceEpoch(0);
 
-    final testDevice = Device(
+    final testDevice = SentryDevice(
       name: 'testDevice',
       family: 'testFamily',
       model: 'testModel',
       modelId: 'testModelId',
       arch: 'testArch',
       batteryLevel: 23,
-      orientation: Orientation.landscape,
+      orientation: SentryOrientation.landscape,
       manufacturer: 'testOEM',
       brand: 'testBrand',
       screenResolution: '123x345',
@@ -40,15 +40,15 @@ void main() {
       bootTime: testBootTime,
       timezone: 'Australia/Melbourne',
     );
-    const testOS = OperatingSystem(name: 'testOS');
+    const testOS = SentryOperatingSystem(name: 'testOS');
     final testRuntimes = [
       const SentryRuntime(name: 'testRT1', version: '1.0'),
       const SentryRuntime(name: 'testRT2', version: '2.3.1'),
     ];
-    const testApp = App(version: '1.2.3');
-    const testBrowser = Browser(version: '12.3.4');
+    const testApp = SentryApp(version: '1.2.3');
+    const testBrowser = SentryBrowser(version: '12.3.4');
 
-    final gpu = Gpu(name: 'Radeon', version: '1');
+    final gpu = SentryGpu(name: 'Radeon', version: '1');
 
     final contexts = Contexts(
       device: testDevice,